requirements.txt
requests==2.25.1
numpy==1.18.5
tqdm==4.60.0
pandas==1.4.2
nltk==3.6.2
streamlit==1.8.1
beautifulsoup4==4.11.1
Pillow==9.1.1
scikit_learn==1.1.1
